Eliminate some unneeded BoxContainer swizzling. Get rid of a gcc warning in GridGeometry.C.
git-svn-id: https://sourceforge.llnl.gov/svn/repos/samrai_project/trunk/MASTER@263 f45145c3-882e-4954-aa44-00140844b9ef

+++ source/SAMRAI/hier/GridGeometry.C
@@ -175,7 +175,7 @@ GridGeometry::GridGeometry(
         ++itr) {
       block_numbers.insert(itr->getBlockId().getBlockValue());
    }
-   d_number_blocks = block_numbers.size();
+   d_number_blocks = static_cast<int>(block_numbers.size());
    d_reduced_connect.resizeArray(d_number_blocks, false);
    d_block_neighbors.resizeArray(d_number_blocks);
